requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
flask==3.0.0
flask-sqlalchemy==3.1.1
schedule==1.2.1
